refactor(uiSelectChoices) only compile template once
Previously the template was compiled, edited, and then recompiled. This
was unecessary as the changes we unrelated to the scope and so could be
made before the link function was run.

This change, moves the DOM transformations to the compile function so
the directive does not have to be compiled manually.

Fixes #1000